Merge interface functionality to the main branch of the app.

+++ app/src/main/java/com/plcoding/weatherapp/data/remote/WeatherDataDto.kt
@@ -13,5 +13,5 @@ data class WeatherDataDto(
     @field:Json(name = "windspeed_10m")
     val windSpeeds: List<Double>,
     @field:Json(name = "relativehumidity_2m")
-    val humidities: List<Int>
+    val humidities: List<Double>
 )
